Add fon9/fmkt/FmktTools.* 升降單位機制、計算漲跌停; Add f9tws/TwsTypes.h 證券商品種類 enum f9tws_SymbKind; Add f9tws/TwsTools.* 根據商品Id編碼規則,判斷: 商品種類、商降單位; Add f9tws 盤中零股; 調整 FileImp 機制; Add make_static_o.sh 協助 gcc 建立 static library; 修正 .c warning;

+++ fon9/CTools.h
@@ -27,7 +27,7 @@ static inline void fon9_SleepMS(useconds_t ms) {
 static inline uint64_t fon9_GetSystemUS() {
    struct timeval tv;
    gettimeofday(&tv, (struct timezone*)NULL);
-   return tv.tv_sec * (uint64_t)1000000 + tv.tv_usec;
+   return (uint64_t)tv.tv_sec * (uint64_t)1000000 + (uint64_t)tv.tv_usec;
 }
 #endif
 fon9_AFTER_INCLUDE_STD;
